+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes
 
+## 2021 Jul 6
+* Refactor Receipt Storage to use AuthorityId
+* Further work on build and send ER, signing state root
+
 ## 2021 Jul 4
 
 * Include executor discovery and initial effort to send ER after block import from Instant Seal